Correct updated terms in Braze
Hi there, I'm one of the technical writers at Braze. We recently made some changes to our navigation and IA in Braze that have affected this documentation. Specifically, we renamed "app groups" to "workspaces" and the **Developer Console** is no longer a page. All subpages of the Developer Console are now their own page (so in this case, the page name should now be **API Keys**). 

This PR fixes those two term changes. Thank you!